requirements.txt
Flask
flask-socketio
PyJWT
pillow


